Add Products schema
