+++ server/dbClients/oracle.js
@@ -61,47 +61,3 @@ async function executeQuery(query) {
 }
 
 module.exports = { executeQuery };
-
-/**
- * create table TEST_USERS ( "id" number generated always as identity, "firstName" varchar2(4000), "lastName" varchar2(4000), "age" number, primary key ("id"));
- * 
- * {
- "query": "mutation createUser($firstName: String!, $lastName: String!, $age: Int!) { createUser(firstName: $firstName, lastName: $lastName, age: $age) }",
- "variables": {
-   "firstName": "Aasif",
-   "lastName": "Rasul",
-   "age": 40
- }
-}
- * 
- * {
-  "query": "{ getUser(id: 1) {id, firstName, lastName, age} }"
-}
- * 
- * 
- * {
-  "query": "{ getUsers {id, firstName, lastName, age} }"
-}
- * 
- * 
- * {
- "query": "mutation UpdateUser($id: ID!, $firstName: String, $lastName: String, $age: Int) { updateUser(id: $id, firstName: $firstName, lastName: $lastName, age: $age) }",
- "variables": {
-   "id": "1",
-   "firstName": "John",
-   "lastName": "Doe",
-   "age": 30
- }
-}
- * 
- * 
- * 
- * {
- "query": "mutation deleteUser($id: ID!) { deleteUser(id: $id) }",
- "variables": {
-   "id": "2"
- }
-}
- * 
- * 
-*/
