+++ server/schema/helper.js
@@ -1,0 +1,12 @@
+const getLimitCond = (dbType, count) => {
+	switch (dbType) {
+		case 'mysql':
+			return `LIMT ${count}`;
+		case 'oracle':
+			return `FETCH FIRST ${count} ROWS ONLY`;
+		default:
+			return '';
+	}
+};
+
+module.exports = { getLimitCond };
